Improve mus/midi event handling
There are differences between MUS and MIDI events. Some of the controller messages are shared, albeit different CC numbers, and some are not.
Special MUS controller messages will have 128 added their CC number to distinguish them.

MUS has an instrument change controller message, whereas MIDI does not.  An event function is required for this purpose.

Channel Aftertouch is not a controller message in MIDI, and should be called more directly.

All Sounds Off is not MIDI level 1 spec. However, the functionality is already there, so it can be implemented.