Handling voices in MIDI events
Most MIDI events have a channel number. In order to find a voice for that channel, requires a search loop iterating through all the voices whether they're playing or not. This is very inefficient and not necessarily very accurate. The solution is to have each channel keep a linked list of voices that are currently playing on that channel. This can be extended to have a linked list of voices that are not playing (a "voice off" list). It's then a simple case of moving from one linked list to another by prepending each voice to the list.